style detection taking shape
