--- sec_parsers/sec_parsers/test_html.py
@@ -1,5 +1,0 @@
-<html>
-<body>
-<div>text LHS div <p> text in paragraph> text RHS div</p></div>
-</body>
-</html>
