+++ sec_parsers/sec_parsers/xml_helper.py
@@ -1,56 +1,37 @@
-# Note, we'll be using 'node' for xml and 'element' for html terminology
-# Note: adding helpers for tree / root / node, so easier for less technical people
-import xml.etree.ElementTree as ET
-
-# lazy so used chatgpt for this
-def is_tree(tree):
-    """
-    Check if the given object is a valid ElementTree instance.
-
-    Parameters:
-    tree (ET.ElementTree): The XML tree to check.
-
-    Returns:
-    bool: True if the object is a valid ElementTree, False otherwise.
-    """
-    try:
-        # Try to get the root element of the tree
-        root = tree.getroot()
-        return True
-    except AttributeError:
-        # The tree object doesn't have a getroot method
-        return False
-    except ET.ParseError:
-        # The tree object is not properly parsed
-        return False
+import webbrowser
+import tempfile
+from lxml import html
 
 
-
-def print_xml_structure(node):
-    if is_tree(tree):
-        root = tree.getroot()
-
-    def indent(level):
-      return "|--" * level
-
-    def print_element(element, level):
-      print(indent(level) + element.tag)
-      for child in element:
-        print_element(child, level + 1)
-
-    print_element(node, 0)
-
-
-def get_text_from_node(node):
+def get_text(element):
+    """Get text from element including tail"""
     text = ''
-    if node.text is not None:
-        text = node.text.strip()
-    for child in node:
-        text += get_text_from_node(child)
-    
+    if element.text:
+        text += element.text
+
+    if element.tail:
+        text += ' ' + element.tail
     return text
 
-def save_xml(tree, file_path):
-    if not is_tree(tree):
-      tree = ET.ElementTree(tree)
-    tree.write(file_path, encoding='utf-8-sig', xml_declaration=True)
+# check includes tail
+def get_all_text(node):
+    """Get all text from element including children. Make include tail"""
+    text = ''.join(node.itertext())
+
+    return text
+
+
+# visualization
+
+def set_background_color(element, color):
+    """Sets the background color for an element."""
+    element.set('style', f'background-color: {color}')
+
+def open_tree(tree):
+    """Opens a lxml tree in a web browser."""
+    with tempfile.NamedTemporaryFile('w', delete=False, suffix='.html', encoding="utf-8-sig") as f:
+        data = html.tostring(tree).decode("utf-8-sig")
+        f.write(data)
+
+    url = 'file://' + f.name
+    webbrowser.open(url)
