+++ stuff.txt
@@ -1,0 +1,2 @@
+# lookup what this is called
+#tree  = etree.fromstring('<html> <body> <div>text1 <p> par1</p> text2 <p>par2</p> text3</div> </body> </html>', parser)
